requirements.txt
construct==2.5.3
Django==2.2.6
django-filter==2.2.0
djangorestframework==3.10.3
drf-nested-routers==0.91
future==0.18.1
Markdown==3.1.1
mysqlclient==1.4.4
pefile==2019.4.18
Pillow==6.2.1
psycopg2-binary==2.8.4
python-dateutil==2.8.0
python-ptrace==0.9.4
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
